feat: Implement API DTOs and custom exceptions for improved request/response handling and error management, alongside refactoring JWT and liking services.

+++ major-project/social-media-be/src/main/java/org/example/socialmediabe/model/Post.java
@@ -1,5 +1,6 @@
 package org.example.socialmediabe.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -21,7 +22,7 @@ import java.util.List;
 public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 1000, nullable = false)
     @NotBlank(message = "Caption cannot be blank")
